fixed crud users and roles

+++ app/Http/Controllers/AdvertController.php
@@ -7,9 +7,21 @@ use App\Models\Company;
 use App\Models\Advert;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\View\View;
+use Illuminate\Http\Request;
+use Illuminate\Http\JsonResponse;
+use RealRashid\SweetAlert\Facades\Alert;
 
 class AdvertController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+        $this->middleware(['permission:advert-list|create-advert|edit-advert|delete-advert'], ['only' => ['index', 'show']]);
+        $this->middleware('permission:create-advert', ['only' => ['create', 'store']]);
+        $this->middleware('permission:update-advert', ['only' => ['edit', 'update']]);
+        $this->middleware('permission:delete-advert', ['only' => 'destroy']);
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -20,6 +32,27 @@ class AdvertController extends Controller
 
         return view('admin.adverts.index', compact('adverts', 'companies'))
             ->with('i', (request()->input('page', 1) - 1) * 5);
+    }
+
+    function search(Request $request)
+    {
+        $searchString = $request->input('search_string');
+
+        $adverts = Advert::with('Company')
+            ->where(function ($query) use ($searchString) {
+                $query->where('title', 'like', '%' . $searchString . '%')
+                    ->orWhere('content', 'like', '%' . $searchString . '%')
+                    ->orWhereHas('company', function ($companyQuery) use ($searchString) {
+                        $companyQuery->where('name', 'like', '%' . $searchString . '%');
+                    });
+            })
+            ->get();
+
+        if (count($adverts)) {
+            return response()->json($adverts);
+        } else {
+            return response()->json(['status' => 'not found']);
+        }
     }
     public function home(): View
     {
@@ -36,7 +69,7 @@ class AdvertController extends Controller
     public function create(): View
     {
         $companies = Company::all();
-        return view('adverts.create', compact('companies'));
+        return view('admin.adverts.create', compact('companies'));
     }
 
     /**
@@ -44,15 +77,18 @@ class AdvertController extends Controller
      */
     public function store(StoreAdvertRequest $request)
     {
-
-
         $validatedData = $request->validated();
 
-        Advert::create($validatedData);
-        // dd($validatedData);
+        $advert = Advert::create($validatedData);
 
-        return redirect()->route('adverts.index')->with('success', 'Advert created successfully.');
+        return response()->json([
+            'message' => 'Advert created successfully',
+            'redirect_url' => route('adverts.index')
+        ]);
     }
+
+
+
 
     /**
      * Display the specified resource.
@@ -72,31 +108,39 @@ class AdvertController extends Controller
         $advert = Advert::findOrFail($id);
         $companies = Company::all();
 
-        return view('adverts.edit', compact('advert', 'companies'));
+        return view('admin.adverts.edit', compact('advert', 'companies'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(StoreAdvertRequest $request, string $id): RedirectResponse
+    public function update(StoreAdvertRequest $request, string $id)
     {
+
         $validatedData = $request->validated();
 
         $advert = Advert::findOrFail($id);
         $advert->update($validatedData);
 
-        return redirect()->route('adverts.index')->with('success', 'Advert updated successfully.');
+        return response()->json([
+            'message' => 'Advert created successfully',
+            'redirect_url' => route('adverts.index')
+        ]);
     }
+
 
     /**
      * Remove the specified resource from storage.
      */
 
-   public function destroy(string $id): RedirectResponse
+    public function destroy(Request $request, string $id): JsonResponse
     {
-        $advert = Advert::findOrFail($id);
-        $advert->delete();
+        if ($request) {
+            $advert = Advert::findOrFail($id);
+            $advert->delete();
+            return response()->json(['status' => 'success', 'message' => 'Success! Advert is deleted']);
+        }
 
-        return redirect()->route('adverts.index')->with('success', 'Advert deleted successfully.');
+        return response()->json(['status' => 'success', 'message' => 'Failed! Unable to delete Advert']);
     }
 }
